% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESIDEMR-package.R
\docType{package}
\name{BESIDEMR-package}
\alias{BESIDEMR-package}
\alias{besidemr}
\alias{BESIDEMR}
\title{BESIDEMR: BayEsian Set IDEntification Mendelian randomization}
\description{
R package to perform BayEsian Set IDEntification Mendelian randomization (BESIDE-MR).
}
\references{
Shapland, C.Y., Zhao Q, Bowden J, Profile-likelihood Bayesian model
averaging for two-sample summary data Mendelian randomization in the presence
of horizontal pleiotropy. Statistics in Medicine, 2022, 41, 6, 1100-1119.
\doi{10.1002/sim.9320}
}
