% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESIDE_MR.R
\name{BMA_MRanalysis}
\alias{BMA_MRanalysis}
\title{BESIDE-MR fitting function}
\usage{
BMA_MRanalysis(
  tau_estimate,
  N_Beta,
  BetaXG,
  BetaYG,
  seBetaXG,
  seBetaYG,
  N_Ins,
  N_Iter,
  Prior,
  tuning_para,
  gen_inits
)
}
\arguments{
\item{tau_estimate}{Use DL estimate (="DL_approx") or Full Bayesian (="Full_Bayes") approach to analyse data.}

\item{N_Beta}{1 parameter (=1) or 2 parameter (=2) model.}

\item{BetaXG}{Effect size for X-G association}

\item{BetaYG}{Effect size for Y-G association}

\item{seBetaXG}{Standard error for X-G association}

\item{seBetaYG}{Standard error for Y-G association}

\item{N_Ins}{Number of genetic instruments}

\item{N_Iter}{Number of iterations}

\item{Prior}{Hyper parameter for the prior distribution;
- For 1 parameter model
   1. $hyper_Beta_mean and $hyper_Beta_sd to specify mean and standard deviation for
        the normally distributed beta
   2. $hyper_Prec_shape and $hyper_Prec_rate to specify shape and rate for
        the gamma distribution of precision ("Full_Bayes" ONLY)
   3. $Ins_prob assign prior inclusion probability for each instrument
- For 2 parameter model
   1. $hyper_Beta1_mean and $hyper_Beta1_sd to specify mean and standard deviation for the normally
        distributed beta1
   2. $hyper_Beta2_mean and $hyper_Beta2_sd to specify mean and standard deviation for the normally
        distributed beta2
   3. $hyper_Prec1_shape and $hyper_Prec1_rate to specify shape and rate for the gamma distribution
        of precision ("Full_Bayes" ONLY)
   4. $hyper_Prec2_shape and $hyper_Prec2_rate to specify shape and rate for the gamma distribution
        of precision ("Full_Bayes" ONLY)
   5. $Ins1_prob assign prior inclusion probability for each instrument in set 1
   6. $Ins2_prob assign prior inclusion probability for each instrument in set 2}

\item{tuning_para}{Tuning parameter to ensure sufficient acceptance rate (recommended between 0.25 - 0.45)
- For 1 parameter model
   1. $Beta for beta
   2. $Prec_LL, $Prec_UL and $Prec_gap for the upper, lower and gap for the precision respectively,
        this ensures symmetry between the new and the old value ("Full_Bayes" ONLY)
- For 2 parameter model
   1. $Beta1 for beta1
   2. $Beta2 for beta2
   3. $Prec1_LL, $Prec1_UL and $Prec1_gap for the upper, lower and gap for the precision1 respectively,
        ("Full_Bayes" ONLY)
   4. $Prec2_LL, $Prec2_UL and $Prec2_gap for the upper, lower and gap for the precision2 respectively,
        ("Full_Bayes" ONLY)}

\item{gen_inits}{Initial values to start the iterations}
}
\value{
An object of class \code{"beside"} containing the following components:\describe{
\item{\code{S}}{A matrix giving the results.}
\item{\code{accept_rate}}{acceptance rate.}
}
}
\description{
Fits BESIDE-MR model. NOTE: beta is the estimated causal effect and tau is the heterogeneity variance modelled as precision (i.e. 1/tau)
}
\examples{

Prior choice for beta, tau and inclusion of instruments
Ins_prior<-rep(0.5, L)
Prior_DL<-list(hyper_Beta_mean=0, hyper_Beta_sd=1, Ins_prob=Ins_prior)
Prior_gamma<-list(hyper_Beta_mean=0, hyper_Beta_sd=1, hyper_Prec_shape=2, hyper_Prec_rate=0.00005, Ins_prob=Ins_prior)

Tuning parameter for beta
H_DL<-list(Beta=0.05)
H_gamma<-list(Beta=0.05, Prec_LL=0, Prec_UL=1000000, Prec_gap=150000)

Generate initial values
gen_inits_DL<-list(Beta=rnorm(1,0,10), Ins_L=randomS.initial.LI(rep(0,L), L,Ins_prior))
gen_inits_gamma<-list(Beta=rnorm(1,0,10), UBPrec=1000000, LBPrec=0, Ins_L=randomS.initial.LI(rep(0,L), L,Ins_prior))

M-H algorithm
res_DL<-BMA_MRanalysis("DL_approx", data$BetaXG,data$BetaYG,data$seBetaXG,data$seBetaYG, L, nIter, Prior_DL, H_DL, gen_inits_DL)
res_gamma<-BMA_MRanalysis("Full_Bayes", data$BetaXG,data$BetaYG,data$seBetaXG,data$seBetaYG, L, nIter, Prior_gamma, H_gamma, gen_inits_gamma)
}
\references{
Shapland, C.Y., et al., Profile-likelihood Bayesian model averaging for two-sample summary data Mendelian randomization in the presence of horizontal pleiotropy.
}
\author{
Chin Yang Shapland; Jack Bowden.
}
