% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESIDE_MR.R
\name{randomS.initial.LI}
\alias{randomS.initial.LI}
\title{Generate initial model space function}
\usage{
randomS.initial.LI(L, ins_prior)
}
\arguments{
\item{L}{Number of instruments.}

\item{ins_prior}{prior for inclusion probability for each instrument.}
}
\value{
numeric sequence of L length as the indicator variable to initialise in \code{BMA_MRanalysis()}.
}
\description{
Generate random initial model space that doesn't produce empty and 1 variable model space.
}
\examples{

Ins_prior<-rep(0.5, L)
randomS.initial.LI(L,Ins_prior)
}
