% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESIDE_MR.R
\name{BMA_MRanalysis}
\alias{BMA_MRanalysis}
\title{BESIDE-MR fitting function}
\usage{
BMA_MRanalysis(
  tau_estimate,
  N_Beta,
  BetaXG,
  BetaYG,
  seBetaXG,
  seBetaYG,
  N_Ins,
  N_Iter,
  Prior,
  tuning_para,
  gen_inits
)
}
\arguments{
\item{tau_estimate}{Use DL estimate (="DL_approx") or Full Bayesian (="Full_Bayes") approach to analyse data.}

\item{N_Beta}{1 parameter (=1) or 2 parameter (=2) model.}

\item{BetaXG}{Effect size for X-G association}

\item{BetaYG}{Effect size for Y-G association}

\item{seBetaXG}{Standard error for X-G association}

\item{seBetaYG}{Standard error for Y-G association}

\item{N_Ins}{Number of genetic instruments}

\item{N_Iter}{Number of iterations}

\item{Prior}{A list of hyper parameter for the prior distribution;\describe{
\item{\code{$hyper_Beta_mean},\code{$hyper_Beta_sd}}{ONE PARAMETER MODEL: to specify mean (\code{_mean}) and
standard deviation (\code{_sd}) respectively for the normally distributed \eqn{\beta}.}
\item{\code{$hyper_Prec_shape}, \code{$hyper_Prec_rate}}{ONE PARAMETER MODEL with "Full_Bayes": to specify
shape (\code{_shape}) and rate (\code{_rate}) for the gamma distribution of precision.}
\item{\code{$Ins_prob}}{ONE PARAMETER MODEL: assign prior inclusion probability for each instrument.}
\item{\code{$hyper_Beta1_mean},\code{$hyper_Beta1_sd},\code{$hyper_Beta2_mean},\code{$hyper_Beta2_sd}}{ONE PARAMETER
MODEL: to specify mean and standard deviation for the normally distributed \eqn{\beta_1} and \eqn{\beta_2} respectively.}
\item{\code{$hyper_Prec1_shape}, \code{$hyper_Prec1_rate}, \code{$hyper_Prec2_shape}, \code{$hyper_Prec2_rate}}{TWO
PARAMETER MODEL with "Full_Bayes": to specify shape and rate for the gamma distribution of precision.}
\item{\code{$Ins1_prob}, \code{$Ins2_prob}}{TWO PARAMETER MODEL: assign prior inclusion probability for each instrument in set 1 or set 2
repectively.}
}}

\item{tuning_para}{Tuning parameter to ensure sufficient acceptance rate (recommended between 0.25 - 0.45);
\describe{
\item{\code{$Beta}}{ONE PARAMETER MODEL: for \eqn{\beta}.}
\item{\code{$Prec_LL}, \code{$Prec_UL}, \code{$Prec_gap}}{ONE PARAMETER MODEL with "Full_Bayes": for the upper (\code{_UL}),
lower (\code{_LL}) and gap (\code{_gap}) of the precision, this ensures symmetry between the new and the old value.}
\item{\code{$Beta1}, \code{$Beta2}}{TWO PARAMETER MODEL: for \eqn{\beta_1} and \eqn{\beta_2} respectively.}
\item{\code{$Prec1_LL}, \code{$Prec1_UL}, \code{$Prec1_gap}, \code{$Prec2_LL}, \code{$Prec2_UL}, \code{$Prec2_gap}}{TWO
PARAMETER MODEL with "Full_Bayes": for the upper, lower and gap of the precision1 and precision2 respectively, this
ensures symmetry between the new and the old value.}
}}

\item{gen_inits}{Initial values to start the iterations; \describe{
\item{\code{$Beta}}{ONE PARAMETER MODEL: for \eqn{\beta}.}
\item{\code{$UBPrec}, \code{$LBPrec}}{ONE PARAMETER MODEL with "Full_Bayes": for upper and lower limit of initial
value of precision respectively.}
\item{\code{$Ins_L}}{ONE PARAMETER MODEL: use \code{randomS.initial.LI()} to generate random initial model space.}
\item{\code{$Beta1}, \code{$Beta2}}{TWO PARAMETER MODEL: for \eqn{\beta_1} and \eqn{\beta_2} respectively.}
\item{\code{$UBPrec1}, \code{$LBPrec1}, \code{$UBPrec2}, \code{$LBPrec2}}{TWO
PARAMETER MODEL with "Full_Bayes": or upper and lower limit of initial value of precision1 and precision2 respectively.}
\item{\code{$Ins1_L}, \code{$Ins2_L}}{TWO PARAMETER MODEL: use randomS.initial.LI() generate random initial model space.}
}}
}
\value{
An object of class \code{"beside"} containing the following components:\describe{
\item{\code{S}}{A matrix giving the results. ONE PARAMETER MODEL: columns gives sampled values of
\eqn{\beta, \tau^2, I_1,...,I_L} for all iterations (rows). TWO PARAMETER MODEL: columns gives sampled values of
\eqn{\beta_1, \beta_2, \tau_1^2, \tau_2^2, I_{11},...,I_{1L}, I_{21},...,I_{2L}} for all iterations (rows).}
\item{\code{accept_rate}}{acceptance rate for each of the parameter from \code{S}.}
}
}
\description{
Fits BESIDE-MR model. NOTE: \eqn{\beta} is the estimated causal effect and \eqn{\tau} is the heterogeneity variance modelled as precision (i.e. \eqn{1/ \tau})
}
\examples{

#load data
data(AMD_HDL)

#Prior choice for beta, tau and inclusion of instruments
L<-nrow(AMD_HDL)
Ins_prior<-rep(0.5, L)
Prior_DL<-list(hyper_Beta_mean=0, hyper_Beta_sd=1, Ins_prob=Ins_prior)
Prior_gamma<-list(hyper_Beta_mean=0, hyper_Beta_sd=1, hyper_Prec_shape=2, hyper_Prec_rate=0.00005, Ins_prob=Ins_prior)

#Tuning parameter for beta
H_DL<-list(Beta=0.05)
H_gamma<-list(Beta=0.05, Prec_LL=0, Prec_UL=1000000, Prec_gap=150000)

#Generate initial values
gen_inits_DL<-list(Beta=rnorm(1,0,10), Ins_L=randomS.initial.LI(L,Ins_prior))
gen_inits_gamma<-list(Beta=rnorm(1,0,10), UBPrec=1000000, LBPrec=0, Ins_L=randomS.initial.LI(L,Ins_prior))

#M-H algorithm
#res_DL<-BMA_MRanalysis("DL_approx", AMD_HDL$BetaXG,AMD_HDL$BetaYG,AMD_HDLseBetaXG,AMD_HDLseBetaYG, L, nIter, Prior_DL, H_DL, gen_inits_DL)
#res_gamma<-BMA_MRanalysis("Full_Bayes", AMD_HDL$BetaXG,AMD_HDLBetaYG,AMD_HDLseBetaXG,AMD_HDLseBetaYG, L, nIter, Prior_gamma, H_gamma, gen_inits_gamma)
}
\references{
Shapland, C.Y., et al., Profile-likelihood Bayesian model averaging for two-sample summary data Mendelian randomization in the presence of horizontal pleiotropy.
}
\author{
Chin Yang Shapland; Jack Bowden.
}
